requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.0.1
pytest==8.4.1
mcp>=1.0.0
httpx==0.28.1
